Carpet: Convert from/to CCTK_INT and from/to CCTK_REAL
CCTK_INT from/to int and CCTK_REAL from/to double where necessary.

darcs-hash:20060613170310-dae7b-de7744ae2c53e0278d9a687ecdadc5634ecc63a7.gz

+++ Carpet/Carpet/src/Storage.cc
@@ -357,7 +357,8 @@ namespace Carpet {
             CCTK_VWarn (1, __LINE__, __FILE__, CCTK_THORNSTRING,
                         "There are not enough time levels for the desired temporal prolongation order in the grid function group \"%s\".  With Carpet::prolongation_order_time=%d, you need at least %d time levels.",
                         groupname,
-                        prolongation_order_time, prolongation_order_time+1);
+                        (int)prolongation_order_time,
+                        (int)(prolongation_order_time+1));
             free (groupname);
           }
         }
